Refactor: VAC패턴 적용
