+++ src/todo/TodoFormView.jsx
@@ -1,13 +1,6 @@
 import React from 'react';
 
-export default function NewTodoForm({
-  currentTodo, onChangeTodo, onSubmit,
-}) {
-  const handleChangeTodo = (event) => {
-    const { value } = event.target;
-    onChangeTodo(value);
-  };
-
+export default function TodoFormView({ currentTodo, handleChangeTodo, onSubmit }) {
   return (
     <>
       <input placeholder="할 일을 입력해 주세요" onChange={handleChangeTodo} value={currentTodo} />
